refactor(requestmanager): reorganize for clarity
refactor requestmanager to clearly designate responsibilities in actor pattern
